Committing the project
This was a toy project to be used as a sort of proof-of-concept. As
such, I might not add to it too much, save doing some cleanup to the
code.

My python syntax is not very pythonic, so if anyone has any pointers,
please let me know how to improve this code.